requirements.txt
opencv-python
keras
scikit-image
pillow
numpy
tensorflow